requirements.txt
-i https://pypi.org/simple
numpy==1.19.5
setuptools==66
wheel==0.38.4
stable-baselines3[extra]==1.7.0
sb3-contrib==1.7.0
contourpy
cycler
fonttools
kiwisolver
matplotlib
packaging
pillow
pyparsing
python-dateutil
six
torch
tqdm
typing-extensions
fastapi
uvicorn[standard]
mlflow
gym
networkx
pyyaml